feat: export components for gen test QR
