+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "anon-aadhaar",
-  "version": "2.0.0",
+  "version": "2.0.1",
   "description": "A set of tools to anonymously prove your identity",
   "main": "index.js",
   "private": true,
