+++ packages/circuits/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@anon-aadhaar/circuits",
-  "version": "2.0.0",
+  "version": "2.0.1",
   "main": "index.js",
   "license": "MIT",
   "scripts": {
