+++ packages/contracts/src/AnonAadhaar.sol
@@ -26,6 +26,7 @@ contract AnonAadhaar is IAnonAadhaar {
     /// @param nullifier: Nullifier for the users Aadhaar.
     /// @param timestamp: Timestamp of when the QR code was signed.
     /// @param signal: Signal committed while genereting the proof.
+    /// @param revealArray: Array of the values used as input for the proof generation (equal to [0, 0, 0, 0] if no field reveal were asked).
     /// @param groth16Proof: SNARK Groth16 proof.
     /// @return Verified bool
     function verifyAnonAadhaarProof(
