+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@anon-aadhaar/core",
-  "version": "2.0.0",
+  "version": "2.0.1",
   "main": "./dist/index.js",
   "types": "./src/index.ts",
   "license": "MIT",
