+++ packages/core/src/index.ts
@@ -4,3 +4,4 @@ export * from './utils'
 export * from './constants'
 export * from './generateArgs'
 export * from './hash'
+export * from './qrGenerator'
