+++ packages/react/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@anon-aadhaar/react",
-  "version": "2.0.0",
+  "version": "2.0.1",
   "description": "Frontend component kit for Anon Aadhaar protocol",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
